+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### 2.15.0
 * `BackupManager::index()` and `BackupManager::rotate()` methods have been deprecated and will be removed in a future
   release. At this point, the entire `BackupManager` class is considered deprecated and will be removed;
-* the `DatabaseBackup.chmod` configuration is deprecated and will be removed in a later release;
+* the `BackupExport::rotate()` method has been deprecated and will be removed in a future release;
+* the `DatabaseBackup.chmod` configuration is deprecated and will be removed in a future release;
 * the `IndexCommand` had been deprecated and was removed;
 * the `--rotate` option of `ExportCommand` had been deprecated and was removed;
 * the magic method `AbstractBackupUtility::__get()` had been deprecated and has been removed;